<feat>implemented check turning on lamp, charging from freezer and charging from battery
