+++ Assets/Scripts/TutorialChecks.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3aa6ba4ce3f73e46a089f4b5c450846
+guid: 4fbca82927ed1d84fa611116b0d07aae
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
