--- Assets/Scripts/TutorialScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e0dfff42fc202b6408ae7b5713446048
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
