--- Assets/Scripts/TutorialScripts/CheckMovement.cs
@@ -1,24 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CheckMovement : MonoBehaviour {
-
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (other.name == "Player")
-        {
-            Debug.Log("Player entered " + this.name);
-        }
-    }
-}
